MH-12738: Adding the TieredStorageAssetManager and TieredStorageAssetManagerJobProducer.
TieredStorageAssetManager is an asset manager with the concept of multiple RemoteAssetStores.  It can move Snapshots
between them based on mediapackage id, version, and date information.  Snapshots moved between two remote asset stores
must be buffered locally first, at least in this implementation.

TieredStorageAssetManagerJobProducer is what enables the above to function asynchronously.  The operations are
synchronous, but this class creates Jobs in the service registry to listen and handle the various events.