Make starting containers also work
Previously the retry mechanism would work in case opensearch would start
as a service (and of course after opencast). With this change it also
works with opensearch in a container. In case a container is started,
an IOException will be thrown with a SocketException as the cause, so this
commit adds a catch for this case and handles it.